Added features/panel/styles.js
