+++ src/features/panel/styles.js
@@ -1,0 +1,19 @@
+const styles = {
+  gameBox: {
+    background: '#f6f6f6',
+  },
+  moves: {
+    height: 300,
+  },
+  buttons: {
+    display: 'flex',
+    justifyContent: 'center',
+    background: '#f0f0f0'
+  },
+  timer: {
+    display: 'flex',
+    justifyContent: 'center',
+  },
+};
+
+export default styles;
